+++ app/src/main/java/org/android/go/sopt/data/local/Dog.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.data
+package org.android.go.sopt.data.local
 
 import androidx.annotation.DrawableRes
