+++ app/src/main/java/org/android/go/sopt/data/local/User.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.data
+package org.android.go.sopt.data.local
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
